Registered new domain cftoolsuite.org; change groupId and packaging to suit
