+++ src/main/java/org/cftoolsuite/actuator/runtime/metadata/CycloneDxInfoContributor.java
@@ -1,4 +1,4 @@
-package io.github.cftoolsuite.actuator.runtime.metadata;
+package org.cftoolsuite.actuator.runtime.metadata;
 
 import java.util.List;
 import java.util.stream.Collectors;
