+++ src/main/java/org/cftoolsuite/actuator/runtime/metadata/PomEndpoint.java
@@ -1,21 +1,19 @@
-package io.github.cftoolsuite.actuator.runtime.metadata;
-
-import java.util.Set;
+package org.cftoolsuite.actuator.runtime.metadata;
 
 import org.springframework.boot.actuate.endpoint.annotation.Endpoint;
 import org.springframework.boot.actuate.endpoint.annotation.ReadOperation;
 
-@Endpoint(id = "jars")
-public class JarsEndpoint {
+@Endpoint(id = "pom")
+public class PomEndpoint {
 
     private final ArtifactsService artifactsService;
 
-    public JarsEndpoint(ArtifactsService artifactsService) {
+    public PomEndpoint(ArtifactsService artifactsService) {
         this.artifactsService = artifactsService;
     }
 
     @ReadOperation
-    public Set<String> getJars() {
-        return artifactsService.findJars();
+    public String getPom() {
+        return artifactsService.findPom();
     }
 }
