+++ src/main/java/org/cftoolsuite/actuator/runtime/metadata/RuntimeMetadataAutoConfiguration.java
@@ -1,4 +1,4 @@
-package io.github.cftoolsuite.actuator.runtime.metadata;
+package org.cftoolsuite.actuator.runtime.metadata;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.actuate.autoconfigure.info.ConditionalOnEnabledInfoContributor;
